Add files via upload
This query creates the customers table to store customer information. Each row represents a customer with a unique customer_id that auto-increments, the first and last names, email, city, state, and the date the customer signed up.

Purpose:
The purpose of this query is to set up the database structure for storing customer data, which will be used in later analysis queries.